+++ runtime/devnet/src/pallets/mod.rs
@@ -9,3 +9,6 @@ pub use timestamp::*;
 
 pub mod balances;
 pub use balances::*;
+
+pub mod transaction_payment;
+pub use transaction_payment::*;
